requirements.txt
Flask
python-dotenv
cachetools

dataclasses_json
aiohttp
tiktoken
tqdm
asyncio




## slack bot
slack_sdk
